refactor: keep coverage entrypoint within size limit
